Moved helpers from todolistutils to todolistconverters.
